refactor: add additional component-wrapper to fix ErrorBoundary behaviour
